requirements.txt
#pdftoscv用のライブラリ
spire.pdf
pandas
#スクレイピング用のライブラリ
chromedriver-binary==128.*
selenium
python-dotenv